requirements.txt
requests==2.31.0
ccxt==4.2.19
pandas==2.2.1
numpy==1.26.4
pyyaml==6.0.1
python-dotenv==1.0.1
ta==0.11.0
backtrader==1.9.78.123
matplotlib==3.8.3
jupyter==1.0.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
backtrader
